Add files via upload
Accurate segmentation of clients is crucial in the success of any business enterprise. It serves the dual purpose of helping the business deal with the client based on their types and also in digital marketing in creating lookalikes. This study is about the binary classification of loan seeking bank customers based on whether they are likely to defualt or not.

When a customer seeks for a loan, banks and other credit providers should use statistical models to determine whether or not to grant the loan based on the likelihood of the loan being repaid or not. The factors involved in determining this likelihood are complex, and extensive statistical analysis and modeling are required to predict the outcome for each individual case. This model predicts loan repayment or default based on the data provided.